Added logic for sourcing non-existent username

+++ assets/warden-source.sh
@@ -8,13 +8,18 @@ echo '  \ V  V / (_| | | | (_| |  __/ | | |'
 echo '   \_/\_/ \__,_|_|  \__,_|\___|_| |_|  Source'
 echo ''
 echo ''
-echo -n "Username: "
+echo -n "Enter a username: "
 read username
-gpg --output ~/opt/warden/$username-creds.sh --decrypt ~/opt/warden/$username-creds.sh.gpg
-wait
-source ~/opt/warden/$username-creds.sh
-wait
-rm -rf ~/opt/warden/$username-creds.sh
-clear
-echo "*** Azure credentials successfully sourced ***"
-echo ""
+
+if [ ! -f ~/opt/warden/$username-creds.sh.gpg ]; then
+    echo "Username not found!"
+else
+  gpg --output ~/opt/warden/$username-creds.sh --decrypt ~/opt/warden/$username-creds.sh.gpg
+  wait
+  source ~/opt/warden/$username-creds.sh
+  wait
+  rm -rf ~/opt/warden/$username-creds.sh
+  clear
+  echo "*** Azure credentials successfully sourced ***"
+  echo ""
+fi
